+++ src/main/java/com/example/hangar/service/PersonaService.java
@@ -8,11 +8,11 @@ public interface PersonaService {
 
     List<Persona> findAll();
 
-    Persona findById(Long id);
+    Persona findById(Integer id);
 
     Persona save(Persona entity);
 
-    void delete(Long id);
+    void delete(Integer id);
 
-    String checkDeletionConstraints(Long id);
+    String checkDeletionConstraints(Integer id);
 }
